Add a simple watchdog python script.
Running this will invoke krad_radio_daemon and keep it running,
restarting it whenever it crashes or is killed. Useful for
keeping streams going. Stop the watchdog itself to kill the daemon.

Very basic at this point; just restarts in a loop. Next useful
bit would be to take a set of krad_radio commands to reconfigure
the daemon after it starts.