Use logging functions instead of printlines
A long time ago, stdout was freopened to handle logging. This does
not allow for such things as sending logs over a socket easily. Now
some logging functions exist, and must be used strictly because we
close stdout/stdin/stderr upon daemonizing, and writing to a closed
file is like walking into a closed door.